core/test/run-trace: Fix type in testing struct
A mock cpu thread structure is used for testing the skiboot trace
buffers. This contains the parts of the actual structure that are needed
for the tests. Within the mock structure the type used for servo_no is
different from the real structure. Change the type so it is consistent.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>