ccan: Add CCAN heap source
We would like to be able to use dump_trace to dump multiple trace
buffers at a time. The entries should be displayed in timestamp order.
As each buffer is already ordered on timestamp, a k-way merge is an
efficient method to sort the buffers together by timestamp. A heap can
be used to implement a k-way merge. As CCAN is already included in
Skiboot, use the CCAN heap. Add the source for heap.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>
[stewart: ccan/heap: Make test run quieter]
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>