pci: Try harder to add meaningful ibm,loc-code
We keep the existing logic of looking to the parent for the slot-label or
slot-location-code, but we add logic to (if all that fails) we look
directly for the slot-location-code (as this should give us the correct
loc code for things directly under the PHB), and otherwise we just look
for a loc-code.

The applicable bit of PAPR here is:

  R1–12.1–1. Each instance of a hardware entity (FRU) has a platform
  unique location code and any node in the OF
  device tree that describes a part of a hardware entity must include the
  “ibm,loc-code” property with a
  value that represents the location code for that hardware entity.

which we weren't really fully obeying at any recent (ever?) point in
time. Now we should do okay, at least for PCI.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>