hw/phb4: Make pci-tracing print at PR_NOTICE
When pci-tracing is enabled we print each trace status message and the
final trace status at PR_ERROR. The final status messages are similar to
those printed when we fail to train in the non-pci-tracing path and this
has resulted in spurious op-test failures.

This patch reduces the log-level of the tracing message to PR_NOTICE so
they're not accidently interpreted as actual error messages. PR_NOTICE
messages are still printed to the console during boot.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>