core/fast-reboot: Add im-feeling-lucky option
Fast reboot gets disabled for a number of reasons e.g. the availability
of nvlink. However this doesn't actually affect the ability to perform fast
reboot if no nvlink device is actually present.

Add a nvram option for fast-reset where if it's set to
"im-feeling-lucky" then perform the fast-reboot irrespective of if it's
previously been disabled.

Signed-off-by: Suraj Jitindar Singh <sjitindarsingh@gmail.com>
Acked-by: Russell Currey <ruscur@russell.cc>
[stewart: update nvram_query_eq to nvram_query_eq_dangerous]
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>