npu2: Clear fence state for a brick being reset
[ Upstream commit d496bb141c978a6dc8a106b3d92e5fc1ad0f8663 ]

Resetting a GPU before resetting an NVLink leads to occasional HMIs
which fence some bricks and prevent the "reset_ntl" procedure from
succeeding at the "reset_ntl_release" step - the host system requires
reboot; there may be other cases like this as well.

This adds clearing of the fence bit in NPU.MISC.FENCE_STATE for
the NVLink which we are about to reset.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>