hw/phb4: Make phb4_training_trace() more general
phb4_training_trace() is used to monitor the Link Training Status
State Machine (LTSSM) of the PHB's data link layer. Currently it is only
used to observe the LTSSM while bringing up the link, but sometimes it's
useful to see what's occurring in other situations (e.g. link disable, or
secondary bus reset). This patch renames it to phb4_link_trace() and
allows the target LTSSM state and a flexible timeout to help in these
situations.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>