Remove unused OPAL_PCI_EEH_FREEZE_STATUS2
This call was introduced all the way back at the end of 2012, before
OPAL was public. The #define for the OPAL call was introduced to the
Linux kernel in June 2013, and the call was never used in any kernel
tree ever (as far as we can find).

Thus, it's quite safe to remove this completely unused and completely
untested OPAL call.

Fixes: https://github.com/open-power/skiboot/issues/99
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>