Remove unused OPAL_GET_XIVE_SOURCE
While this call was technically implemented by skiboot, no code has ever called
it, and it was only ever implemented for the p7ioc-phb back-end (i.e. POWER7).
Since this call was unused in Linux, and that  POWER7 with OPAL was only ever
available internally, so it should be safe to remove the call.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>