sparse: fix OPAL API function argument test
This is based on a patch from Cédric from way back in 2015 (probably
around the time we split opal.h into opal-api.h and opal-internal.h)
that for some reason never got merged.

It means that false warnings spat out of sparse due to our crazy-ass
macro usage get silenced.

It fixes warnings such as

    "Using plain integer as NULL pointer"

when the prototype of the function is tested with the __test_args#
defines.

Suggested-by: Cédric Le Goater <clg@fr.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>