libflash: Fix broken continuations
Some of the libflash debug messages don't print a newlines at the end of
the line and assume that the next print will be contigious with the
last. This isn't true in skiboot since log messages are prefixed with a
timestamp. This results in funny looking output such as:

LIBFLASH: Verifying...
LIBFLASH:   reading page 0x01963000..0x01964000...[3.084846885,7]  same !
LIBFLASH:   reading page 0x01964000..0x01965000...[3.086164489,7]  same !

Fix this by moving the "same !" debug message to a new line with the
prefix "LIBFLASH:   ..." to indicate it's a continuation of the last
statement.

First reported in https://github.com/open-power/skiboot/issues/51

Reported-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>