hw/sbe-p9: Fix multi-line log messages
When sending messages to the SBE we log the message using a multi-line
log message that looks like this:

[   96.390873752,8] SBE: Message queued [chip id = 0x0]:
         Reg0 : 000002010054d401
         Reg1 : 0000000000030d40
         Reg2 : 0000000000000000
         Reg3 : 0000000000000000

The lack of a common prefix makes the log messages annoying to deal with
since you can just grep for SBE: to get all the SBE related messages,
and you can't use grep -v to remove them. There's no real benifit to
squashing all this into a single prlog() call, so use a for loop to
print the registers. With this patch the output is:

[   93.253511545,8] SBE: Message queued [chip id = 0x0]:
[   93.253512343,8] SBE:     Reg0 : 000002010059d401
[   93.253513167,8] SBE:     Reg1 : 0000000000030d40
[   93.253513894,8] SBE:     Reg2 : 0000000000000000
[   93.253514627,8] SBE:     Reg3 : 0000000000000000

Cc: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Acked-by: Stewart Smith <stewart@linux.ibm.com>
Reviewed-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>